# Major filtering optimisation + additional features