--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "amsifyselect",
-  "version": "1.2.0",
+  "version": "1.3.1",
   "dependencies": {
     "jquery": ">=3.4.0"
   },