Hi,

**jquery.amsify.select** is a great product and I've decided to improve it even a bit more.

The main reason for this PR is the major optimisation of the filtering functionality. Here is the test results (you can recreate it yourself):
- Browser: Chrome 98.0.4758.102 on macOS Version 10.15.7
- Data: _&lt;select&gt;_ with the following values: 1 to 9998
- Input: 99 (two keystrokes)
   - Filtered results rendering performance before optimisation:
      - 9 (1st keystroke) - 3303ms
      - 99 (2nd keystroke) - 247ms
   - Filtered results rendering performance after optimisation:
      - 9 (1st keystroke) - 95ms
      - 9 (2nd keystroke) - 37ms

The filtering changes were additionally tested using the Firefox browser.
Additionally, the _optgroup_ and _multiple_ _&lt;select&gt; types were tested.

Additional features added:
- _Amsify Selection Area_ now inherits the ID from the original _&lt;select&gt;_ element - appending **_amsify_selection_area** to the inherited value
- The control buttons can now be hidden (refer to the **5. Hide Buttons** section of the **README.md** file
- Updated **README.md** file
- Minor refactoring

I hope that you find these improvements useful.

Thank for your time,
- Andrey